Rework monitor configuration switching
Executing the monitor configuration scripts now writes the configuration
name to ```~/.config/monitorlayout```. The xinitrc, zprofile and zschenv
scripts are then responsible for calling xrandr, starting the X11 server
with the correct dpi and setting dpi related enviroment variables
respectively. Note that since this happens on login, changing the
configuration requires logging in and out to take effect.